+++ demo/shootguy/game.rb
@@ -1,0 +1,5 @@
+require 'yeah'
+
+class ShootguyGame < Yeah::Game
+  # first_map :training # if unspecified, picks first file in /maps
+end