+++ lib/yeah.rb
@@ -1,5 +1,5 @@
 module Yeah
-  VERSION = '0.2.0'
+  VERSION = '0.2.1'
 end
 
 requires = %i(vector color surface rectangle entity map desktop game)