+++ lib/yeah/color.rb
@@ -1,22 +1,25 @@
+# Color.
 class Yeah::Color
+  # @!attribute rgba_bytes
+  #   @return [Array<(Integer, Integer, Integer, Integer)>] red, green, blue,
+  #     alpha bytes
   attr_accessor :rgba_bytes
 
   class << self
     alias_method :[], :new
   end
 
-  def initialize(*arguments)
-    arguments = [0, 0, 0, 255] if arguments.empty?
+  def initialize(*values)
+    default_values = [0, 0, 0, 255]
+    values += default_values[values.size..-1]
+    @rgba_bytes = values
+  end
 
-    case arguments[0]
-    when Numeric
-      @rgba_bytes = [*arguments]
-    when Array
-      @rgba_bytes = arguments[0]
-    end
+  def inspect
+    "#{self.class.name}[#{rgba_bytes.join(', ')}]"
   end
 
   def ==(other)
-    self.class == other.class && @rgba_bytes == other.rgba_bytes ? true : false
+    self.class == other.class && @rgba_bytes == other.rgba_bytes
   end
 end