+++ lib/yeah/game.rb
@@ -1,12 +1,38 @@
+# Manages entities.
 class Yeah::Game
-  attr_accessor :entities, :screen, :platform, :resolution
+  # @!attribute resolution
+  #   @return [Vector] size of screen
+  # @!attribute screen
+  #   @return [Surface] visual render
+  # @!attribute [r] platform
+  #   @return [Platform] underlying platform bindings
+  # @!attribute entities
+  #   @return [Array] active entities
+  attr_accessor :resolution, :screen, :entities
+  attr_reader :platform
 
   def initialize
-    @resolution = Vector[320, 240]
+    @resolution = Vector[320, 180]
     @screen = Surface.new(@resolution)
     @platform = Desktop.new
     @entities = []
   end
+
+  # Start the game loop.
+  def start
+    platform.each_tick do
+      update
+      draw
+      break if @stopped
+    end
+  end
+
+  # Stop the game loop.
+  def stop
+    @stopped = true
+  end
+
+  protected
 
   def update
     @entities.each { |e| e.update }
@@ -19,12 +45,5 @@ class Yeah::Game
       screen.draw(surface, entity.position) unless surface.nil?
     end
     platform.render(screen)
-  end
-
-  def start
-    platform.each_tick do
-      update
-      draw
-    end
   end
 end