+++ lib/yeah/map.rb
@@ -1,4 +1,7 @@
+# A map of entities for a Game.
 class Yeah::Map
+  # @!attribute background
+  #   @return [Color] background color
   attr_accessor :background
 
   def initialize